+++ packages/core/icon/glyph/star-filled.js
@@ -17,7 +17,7 @@ var _index2 = _interopRequireDefault(_index);
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var StarFilledIcon = function StarFilledIcon(props) {
-  return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><path d="M12 17.514l-3.498 1.928a1 1 0 0 1-1.47-1.037l.68-4.156-2.923-2.986a1 1 0 0 1 .564-1.688l3.997-.609 1.745-3.706a1 1 0 0 1 1.81 0l1.745 3.706 3.997.61a1 1 0 0 1 .564 1.687l-2.923 2.986.68 4.156a1 1 0 0 1-1.47 1.037L12 17.514z" fill="currentColor" fill-rule="evenodd"/></svg>' }, props));
+  return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><path d="M12.072 17.284l-3.905 2.053a1 1 0 0 1-1.451-1.054l.745-4.349-3.159-3.08a1 1 0 0 1 .554-1.705l4.366-.635 1.953-3.956a1 1 0 0 1 1.794 0l1.952 3.956 4.366.635a1 1 0 0 1 .555 1.705l-3.16 3.08.746 4.349a1 1 0 0 1-1.45 1.054l-3.906-2.053z" fill="currentColor" fill-rule="evenodd"/></svg>' }, props));
 };
 StarFilledIcon.displayName = 'StarFilledIcon';
 exports.default = StarFilledIcon;