+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/picker-facade-spec.ts
@@ -188,7 +188,6 @@ describe('Media PickerFacade', () => {
           pickerType,
           getPickerFacadeConfig(stateManager),
           {
-            userAuthProvider,
             uploadParams: { collection: '' },
           },
         );