+++ packages/elements/mention/src/api/PresenceResource.ts
@@ -137,9 +137,7 @@ class PresenceResource extends AbstractPresenceResource {
       credentials: 'include' as 'include',
       body: JSON.stringify(data),
     };
-    return fetch(new Request(this.config.url, options)).then(response =>
-      response.json(),
-    );
+    return fetch(this.config.url, options).then(response => response.json());
   }
 
   private static cleanUrl(url: string): string {