+++ packages/media/media-picker/src/popup/middleware/fetchNextCloudFilesPage.tsx
@@ -12,7 +12,7 @@ export const fetchNextCloudFilesPageMiddleware = (fetcher: Fetcher) => (
   store: Store<State>,
 ) => (next: Dispatch<State>) => (action: Action) => {
   if (isFetchNextCloudFilesPageAction(action)) {
-    const { userAuthProvider } = store.getState();
+    const { userContext } = store.getState();
     const { serviceName, accountId, path } = action;
     const { id: folderId } = path[path.length - 1] || { id: '' };
     const { view } = store.getState();
@@ -20,7 +20,8 @@ export const fetchNextCloudFilesPageMiddleware = (fetcher: Fetcher) => (
     const cursor = view && view.nextCursor;
     const items = (view && view.items) || [];
 
-    userAuthProvider()
+    userContext.config
+      .authProvider()
       .then(auth =>
         fetcher.fetchCloudAccountFolder(
           auth,