+++ packages/core/navigation-next/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/navigation-next
 
+## 3.7.2
+- [patch] Fixing multiple skeleton load states, rendering product and container skeletons [68e49a1](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/68e49a1)
+
 ## 3.7.1
 - [patch] Adds sideEffects: false to allow proper tree shaking [b5d6d04](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b5d6d04)